fix(test): update CredEpochs synthesis for gouroboros 0.205.3
gouroboros 0.205.3 (merged from main, #4356) changed
UpdateCommitteeGovAction.CredEpochs from map[*Credential]uint to
map[*Credential]uint64. syntheticUpdateCommitteeGovAction built the
former, which no longer compiles; ProposedMembers/ProposedMembersByCredential
were already uint64, so this drops the now-unnecessary conversion.

+++ internal/test/conformance/state_manager.go
@@ -1525,17 +1525,17 @@ func syntheticUpdateCommitteeGovAction(
 	proposal *conformance.ProposalState,
 ) *common.UpdateCommitteeGovAction {
 	credEpochs := make(
-		map[*common.Credential]uint,
+		map[*common.Credential]uint64,
 		len(proposal.ProposedMembersByCredential)+len(proposal.ProposedMembers),
 	)
 	for _, expiry := range proposal.ProposedMembersByCredential {
-		credEpochs[new(common.Credential)] = uint(expiry)
+		credEpochs[new(common.Credential)] = expiry
 	}
 	for hash, expiry := range proposal.ProposedMembers {
 		if hasCredentialHash(proposal.ProposedMembersByCredential, hash) {
 			continue
 		}
-		credEpochs[new(common.Credential)] = uint(expiry)
+		credEpochs[new(common.Credential)] = expiry
 	}
 	return &common.UpdateCommitteeGovAction{CredEpochs: credEpochs}
 }
